\name{DWDLargeR-package}
\alias{DWDLargeR-package}
\docType{package}
\title{
Fast Algorithms for Large Scale Generalized Distance Weighted Discrimination
}
\description{
  Solving large scale distance weighted discrimination.
  The main algorithm is a symmetric Gauss-Seidel based alternating direction method 
  of multipliers (ADMM) method.
}
\details{
The package \code{DWDLargeR} contains two main functions:\cr
\code{penaltyParameter}\cr
\code{genDWD}\cr
}
\author{Xin-Yee Lam, J.S. Marron, Defeng Sun, and Kim-Chuan Toh}
\references{
Lam, X.Y., Marron, J.S., Sun, D.F., and Toh, K.C. (2018)
  ``Fast algorithms for large scale generalized distance weighted discrimination", \emph{Journal of Computational and Graphical Statistics}, forthcoming. \cr
  \url{https://arxiv.org/abs/1604.05473}
}
 
\keyword{DWD}
\keyword{high-dimensional}
\keyword{classification}
\keyword{package}

