\name{mushrooms}
\docType{data}
\alias{mushrooms}
\title{Classification data from Audobon Society Field Guide (1981).}
\description{
This data set includes descriptions of hypothetical samples corresponding to 23 
species of gilled mushrooms in the Agaricus and Lepiota Family (pp. 500-525). 
Each species is identified as definitely edible, definitely poisonous, or of 
unknown edibility and not recommended. This latter class was combined with the 
poisonous one. The Guide clearly states that there is no simple rule for determining 
the edibility of a mushroom; no rule like ``leaflets three, let it be'' for Poisonous 
Oak and Ivy.
}
\usage{data(mushrooms)}
\format{List containing a 112x8124 matrix of 8124 training samples with 112 features; and a vector of length 8124 training labels.}

\source{The data could be downloaded from the UCI Machine Learning Repository. 
\url{https://archive.ics.uci.edu/ml/datasets/Mushroom }}

\references{Lichman, M. (2013). UCI Machine Learning Repository [http://archive.ics.uci.edu/ml]. Irvine, CA: University of California, School of Information and Computer Science.}
\keyword{datasets}
